exchange_manager_memory in internal scope
